back-character-forge/src/main/java/com/jawbr/dnd5e/characterforge/controller/RaceController.java
package com.jawbr.dnd5e.characterforge.controller;

import com.jawbr.dnd5e.characterforge.dto.response.race.RaceDTOResponse;
import com.jawbr.dnd5e.characterforge.service.RaceService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/races")
public class RaceController {

    private final RaceService raceService;

    public RaceController(RaceService raceService) {
        this.raceService = raceService;
    }

    @GetMapping
    public RaceDTOResponse findAllRaces() {
        return raceService.findAllRaces();
    }
}
